+++ Source/Core/Core/CMakeLists.txt
@@ -148,10 +148,12 @@ set(SRCS	ActionReplay.cpp
 			IOS/FS/FS.cpp
 			IOS/Network/Config.cpp
 			IOS/Network/ICMPLin.cpp
-			IOS/Network/NWC24Config.cpp
 			IOS/Network/Socket.cpp
 			IOS/Network/Net.cpp
 			IOS/Network/SSL.cpp
+			IOS/Network/KD/NetKDRequest.cpp
+			IOS/Network/KD/NetKDTime.cpp
+			IOS/Network/KD/NWC24Config.cpp
 			IOS/SDIO/SDIOSlot0.cpp
 			IOS/STM/STM.cpp
 			IOS/USB/Common.cpp
