+++ Source/Core/Core/Core.vcxproj
@@ -180,10 +180,12 @@
     <ClCompile Include="IOS\FS\FS.cpp" />
     <ClCompile Include="IOS\Network\Config.cpp" />
     <ClCompile Include="IOS\Network\ICMPLin.cpp" />
-    <ClCompile Include="IOS\Network\NWC24Config.cpp" />
     <ClCompile Include="IOS\Network\Socket.cpp" />
     <ClCompile Include="IOS\Network\Net.cpp" />
     <ClCompile Include="IOS\Network\SSL.cpp" />
+    <ClCompile Include="IOS\Network\KD\NetKDRequest.cpp" />
+    <ClCompile Include="IOS\Network\KD\NetKDTime.cpp" />
+    <ClCompile Include="IOS\Network\KD\NWC24Config.cpp" />
     <ClCompile Include="IOS\SDIO\SDIOSlot0.cpp" />
     <ClCompile Include="IOS\STM\STM.cpp" />
     <ClCompile Include="IOS\USB\Common.cpp" />
@@ -417,11 +419,13 @@
     <ClInclude Include="IOS\FS\FS.h" />
     <ClInclude Include="IOS\Network\Config.h" />
     <ClInclude Include="IOS\Network\ICMPLin.h" />
-    <ClInclude Include="IOS\Network\NWC24Config.h" />
     <ClInclude Include="IOS\Network\ICMP.h" />
     <ClInclude Include="IOS\Network\Socket.h" />
     <ClInclude Include="IOS\Network\Net.h" />
     <ClInclude Include="IOS\Network\SSL.h" />
+    <ClInclude Include="IOS\Network\KD\NetKDRequest.h" />
+    <ClInclude Include="IOS\Network\KD\NetKDTime.h" />
+    <ClInclude Include="IOS\Network\KD\NWC24Config.h" />
     <ClInclude Include="IOS\SDIO\SDIOSlot0.h" />
     <ClInclude Include="IOS\STM\STM.h" />
     <ClInclude Include="IOS\USB\Common.h" />
