+++ Source/Core/Core/IOS/Network/Net.h
@@ -7,13 +7,9 @@
 #include <string>
 
 #include "Common/CommonTypes.h"
-#include "Common/Logging/Log.h"
-#include "Core/HW/EXI/EXI_DeviceIPL.h"
-#include "Core/HW/Memmap.h"
 #include "Core/IOS/Device.h"
 #include "Core/IOS/IPC.h"
 #include "Core/IOS/Network/Config.h"
-#include "Core/IOS/Network/NWC24Config.h"
 
 #ifdef _WIN32
 #include <ws2tcpip.h>
@@ -65,143 +61,6 @@ enum NET_IOCTL
 // TODO: split this up.
 namespace Device
 {
-//////////////////////////////////////////////////////////////////////////
-// KD is the IOS module responsible for implementing WiiConnect24 functionality.
-// It can perform HTTPS downloads, send and receive mail via SMTP, and execute a
-// JavaScript-like language while the Wii is in standby mode.
-class NetKDRequest : public Device
-{
-public:
-  NetKDRequest(u32 device_id, const std::string& device_name);
-
-  virtual ~NetKDRequest();
-
-  IPCCommandResult IOCtl(const IOCtlRequest& request) override;
-
-private:
-  enum
-  {
-    IOCTL_NWC24_SUSPEND_SCHEDULAR = 0x01,
-    IOCTL_NWC24_EXEC_TRY_SUSPEND_SCHEDULAR = 0x02,
-    IOCTL_NWC24_EXEC_RESUME_SCHEDULAR = 0x03,
-    IOCTL_NWC24_KD_GET_TIME_TRIGGERS = 0x04,
-    IOCTL_NWC24_SET_SCHEDULE_SPAN = 0x05,
-    IOCTL_NWC24_STARTUP_SOCKET = 0x06,
-    IOCTL_NWC24_CLEANUP_SOCKET = 0x07,
-    IOCTL_NWC24_LOCK_SOCKET = 0x08,
-    IOCTL_NWC24_UNLOCK_SOCKET = 0x09,
-    IOCTL_NWC24_CHECK_MAIL_NOW = 0x0A,
-    IOCTL_NWC24_SEND_MAIL_NOW = 0x0B,
-    IOCTL_NWC24_RECEIVE_MAIL_NOW = 0x0C,
-    IOCTL_NWC24_SAVE_MAIL_NOW = 0x0D,
-    IOCTL_NWC24_DOWNLOAD_NOW_EX = 0x0E,
-    IOCTL_NWC24_REQUEST_GENERATED_USER_ID = 0x0F,
-    IOCTL_NWC24_REQUEST_REGISTER_USER_ID = 0x10,
-    IOCTL_NWC24_GET_SCHEDULAR_STAT = 0x1E,
-    IOCTL_NWC24_SET_FILTER_MODE = 0x1F,
-    IOCTL_NWC24_SET_DEBUG_MODE = 0x20,
-    IOCTL_NWC24_KD_SET_NEXT_WAKEUP = 0x21,
-    IOCTL_NWC24_SET_SCRIPT_MODE = 0x22,
-    IOCTL_NWC24_REQUEST_SHUTDOWN = 0x28,
-  };
-
-  enum
-  {
-    MODEL_RVT = 0,
-    MODEL_RVV = 0,
-    MODEL_RVL = 1,
-    MODEL_RVD = 2,
-    MODEL_ELSE = 7
-  };
-
-  u8 GetAreaCode(const std::string& area) const;
-  u8 GetHardwareModel(const std::string& model) const;
-
-  s32 NWC24MakeUserID(u64* nwc24_id, u32 hollywood_id, u16 id_ctr, u8 hardware_model, u8 area_code);
-
-  NWC24::NWC24Config config;
-};
-
-//////////////////////////////////////////////////////////////////////////
-class NetKDTime : public Device
-{
-public:
-  NetKDTime(u32 device_id, const std::string& device_name)
-      : Device(device_id, device_name), rtc(), utcdiff()
-  {
-  }
-
-  virtual ~NetKDTime() {}
-  IPCCommandResult IOCtl(const IOCtlRequest& request) override
-  {
-    s32 result = 0;
-    u32 common_result = 0;
-    // TODO Writes stuff to /shared2/nwc24/misc.bin
-    // u32 update_misc = 0;
-
-    switch (request.request)
-    {
-    case IOCTL_NW24_GET_UNIVERSAL_TIME:
-      Memory::Write_U64(GetAdjustedUTC(), request.buffer_out + 4);
-      break;
-
-    case IOCTL_NW24_SET_UNIVERSAL_TIME:
-      SetAdjustedUTC(Memory::Read_U64(request.buffer_in));
-      // update_misc = Memory::Read_U32(request.buffer_in + 8);
-      break;
-
-    case IOCTL_NW24_SET_RTC_COUNTER:
-      rtc = Memory::Read_U32(request.buffer_in);
-      // update_misc = Memory::Read_U32(request.buffer_in + 4);
-      break;
-
-    case IOCTL_NW24_GET_TIME_DIFF:
-      Memory::Write_U64(GetAdjustedUTC() - rtc, request.buffer_out + 4);
-      break;
-
-    case IOCTL_NW24_UNIMPLEMENTED:
-      result = -9;
-      break;
-
-    default:
-      ERROR_LOG(IOS_NET, "%s - unknown IOCtl: %x", GetDeviceName().c_str(), request.request);
-      break;
-    }
-
-    // write return values
-    Memory::Write_U32(common_result, request.buffer_out);
-    return GetDefaultReply(result);
-  }
-
-private:
-  enum
-  {
-    IOCTL_NW24_GET_UNIVERSAL_TIME = 0x14,
-    IOCTL_NW24_SET_UNIVERSAL_TIME = 0x15,
-    IOCTL_NW24_UNIMPLEMENTED = 0x16,
-    IOCTL_NW24_SET_RTC_COUNTER = 0x17,
-    IOCTL_NW24_GET_TIME_DIFF = 0x18,
-  };
-
-  u64 rtc;
-  s64 utcdiff;
-
-  // TODO: depending on CEXIIPL is a hack which I don't feel like
-  // removing because the function itself is pretty hackish;
-  // wait until I re-port my netplay rewrite
-
-  // Returns seconds since Wii epoch
-  // +/- any bias set from IOCTL_NW24_SET_UNIVERSAL_TIME
-  u64 GetAdjustedUTC() const { return CEXIIPL::GetEmulatedTime(CEXIIPL::WII_EPOCH) + utcdiff; }
-  // Store the difference between what the Wii thinks is UTC and
-  // what the host OS thinks
-  void SetAdjustedUTC(u64 wii_utc)
-  {
-    utcdiff = CEXIIPL::GetEmulatedTime(CEXIIPL::WII_EPOCH) - wii_utc;
-  }
-};
-
-//////////////////////////////////////////////////////////////////////////
 class NetIPTop : public Device
 {
 public:
